Fix mypy type errors introduced by module extraction refactoring
- Add TYPE_CHECKING imports and proper types for extracted functions
  (risk_sizing, order_lifecycle, reprice_execution) instead of using
  object/Any parameters
- Extend OrderbookLike protocol with is_stale(), spread_bps(),
  orderbook_imbalance(window_s) and return PriceLevelLike from
  best_bid/best_ask
- Add allowed_order_size() to RiskManagerLike protocol
- Declare _pricing, _post_only, _reprice, _drawdown_stop,
  _level_pof_until type annotations on MarketMakerStrategy
- Use Mapping instead of Dict for covariant function parameters
- Add bool() wraps for no-any-return from getattr comparisons
- Guard None fee_resolver before calling refresh/validate
- Fix variable shadowing in mm_config_pipeline._value_with_rule

Reduces mypy errors from 42 pre-existing to 8 (net -25 fixed beyond
the 28 new ones eliminated).

+++ src/market_maker/strategy_components.py
@@ -287,4 +287,4 @@ def estimate_tick_time_ms(s: Any) -> float:
     vol_per_ms = float(vol_bps) / (window_s * 1000.0)
     if vol_per_ms <= 0:
         return 0.0
-    return tick_bps / vol_per_ms
+    return float(tick_bps / vol_per_ms)
